+++ frontend/src/components/roomPreview/RoomPreview.tsx
@@ -14,7 +14,7 @@ const getImage = async () => {
 };
 
 
-export const ChattingRoom = () =>{
+export const RoomPreview = () =>{
 
   // 에러가 발생하면 react-query update를 하기
   const { isLoading, data, isError } = useQuery({
